RestControllerAdvice를 통한 Exception 처리 완료

+++ src/main/java/travel/travel/service/StoryService.java
@@ -3,6 +3,8 @@ package travel.travel.service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import travel.travel.apiPayload.code.status.ErrorStatus;
+import travel.travel.apiPayload.exception.handler.StoryHandler;
 import travel.travel.domain.Story;
 import travel.travel.domain.Writing;
 import travel.travel.dto.story.StoryRequest;
@@ -21,16 +23,17 @@ public class StoryService {
 
     private final StoryRepository storyRepository;
     private final WritingService writingService;
-    private final WritingRepository writingRepository;
 
-    public Long create(String accessToken, StoryRequest.StoryCreateRequestDTO storyCreateRequestDto) {
+    public Long create(StoryRequest.StoryCreateRequestDTO storyCreateRequestDto) {
         Story story = StoryMapper.toStoryFromStoryCreateRequestDto(storyCreateRequestDto);
         Story savedStory = storyRepository.save(story);
         return savedStory.getId();
     }
 
     public StoryResponse.StoryInfoResponseDTO upload(Long storyId) {
-        Story story = storyRepository.findById(storyId).orElseThrow(() -> new NullPointerException("해당 일기가 없습니다"));
+        Story story = storyRepository.findById(storyId).orElse(null);
+        if (story == null)
+            throw new StoryHandler(ErrorStatus._STORY_NOT_FOUND);
         return StoryMapper.toStoryInfoResponseDto(story);
     }
 
@@ -53,13 +56,17 @@ public class StoryService {
     }
 
     public void update(Long storyId, StoryRequest.StoryUpdateRequestDTO storyUpdateRequestDto) {
-        Story story = storyRepository.findById(storyId).orElseThrow(() -> new NullPointerException("해당 일기가 없습니다"));
+        Story story = storyRepository.findById(storyId).orElse(null);
+        if (story == null)
+            throw new StoryHandler(ErrorStatus._STORY_NOT_FOUND);
         story.update(storyUpdateRequestDto);
         storyRepository.save(story);
     }
 
     public void delete(Long storyId) {
-        Story story = storyRepository.findById(storyId).orElseThrow(() -> new NullPointerException("해당 일기가 없습니다"));
+        Story story = storyRepository.findById(storyId).orElse(null);
+        if (story == null)
+            throw new StoryHandler(ErrorStatus._STORY_NOT_FOUND);
         storyRepository.delete(story);
     }
 }
