+++ src/main/java/travel/travel/service/WritingService.java
@@ -6,6 +6,10 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import travel.travel.apiPayload.code.status.ErrorStatus;
+import travel.travel.apiPayload.exception.handler.AccessTokenHandler;
+import travel.travel.apiPayload.exception.handler.StoryHandler;
+import travel.travel.apiPayload.exception.handler.UserHandler;
 import travel.travel.domain.Story;
 import travel.travel.domain.User;
 import travel.travel.domain.Writing;
@@ -27,11 +31,17 @@ public class WritingService {
     private final JwtService jwtService;
 
     public void save(String accessToken, Long storyId) {
-        String email = jwtService.extractEmail(accessToken).orElseThrow(() -> new IllegalStateException("유효하지 않은 토큰입니다."));
-        User writer = userRepository.findByEmail(email).orElseThrow(() -> new NullPointerException("해당하는 사용자가 없습니다."));
-        Story story = storyRepository.findById(storyId).orElseThrow(() -> new NullPointerException("해당하는 일기가 없습니다"));
+        String email = jwtService.extractEmail(accessToken).orElse(null);
+        if (email == null)
+            throw new AccessTokenHandler(ErrorStatus._ACCESSTOKEN_NOT_VALID);
+        User user = userRepository.findByEmail(email).orElse(null);
+        if (user == null)
+            throw new UserHandler(ErrorStatus._USER_NOT_FOUND);
+        Story story = storyRepository.findById(storyId).orElse(null);
+        if (story == null)
+            throw new StoryHandler(ErrorStatus._STORY_NOT_FOUND);
         Writing writing = Writing.builder()
-                .writer(writer)
+                .writer(user)
                 .story(story)
                 .build();
 
@@ -39,29 +49,43 @@ public class WritingService {
     }
 
     public List<Writing> getByPaging(String accessToken, Integer page_number, Integer page_size) {
-        String email = jwtService.extractEmail(accessToken).orElseThrow(() -> new IllegalStateException("유효하지 않은 토큰입니다."));
-        User writer = userRepository.findByEmail(email).orElseThrow(() -> new NullPointerException("해당하는 사용자가 없습니다."));
+        String email = jwtService.extractEmail(accessToken).orElse(null);
+        if (email == null)
+            throw new AccessTokenHandler(ErrorStatus._ACCESSTOKEN_NOT_VALID);
+        User user = userRepository.findByEmail(email).orElse(null);
+        if (user == null)
+            throw new UserHandler(ErrorStatus._USER_NOT_FOUND);
 
         Pageable pageable = PageRequest.of(page_number - 1, page_size);
-        Page<Writing> pagedWritings = writingRepository.findAllByWriter(writer, pageable);
+        Page<Writing> pagedWritings = writingRepository.findAllByWriter(user, pageable);
 
         return pagedWritings.getContent();
     }
 
     public List<Writing> getAll(String accessToken) {
-        String email = jwtService.extractEmail(accessToken).orElseThrow(() -> new IllegalStateException("유효하지 않은 토큰입니다."));
-        User writer = userRepository.findByEmail(email).orElseThrow(() -> new NullPointerException("해당하는 사용자가 없습니다."));
-        return writingRepository.findAllByWriter(writer);
+        String email = jwtService.extractEmail(accessToken).orElse(null);
+        if (email == null)
+            throw new AccessTokenHandler(ErrorStatus._ACCESSTOKEN_NOT_VALID);
+        User user = userRepository.findByEmail(email).orElse(null);
+        if (user == null)
+            throw new UserHandler(ErrorStatus._USER_NOT_FOUND);
+        return writingRepository.findAllByWriter(user);
     }
 
     public int count(String accessToken) {
-        String email = jwtService.extractEmail(accessToken).orElseThrow(() -> new IllegalStateException("유효하지 않은 토큰입니다."));
-        User writer = userRepository.findByEmail(email).orElseThrow(() -> new NullPointerException("해당하는 사용자가 없습니다."));
-        return writingRepository.findAllByWriter(writer).size();
+        String email = jwtService.extractEmail(accessToken).orElse(null);
+        if (email == null)
+            throw new AccessTokenHandler(ErrorStatus._ACCESSTOKEN_NOT_VALID);
+        User user = userRepository.findByEmail(email).orElse(null);
+        if (user == null)
+            throw new UserHandler(ErrorStatus._USER_NOT_FOUND);
+        return writingRepository.findAllByWriter(user).size();
     }
 
     public void delete(Long storyId) {
-        Story story = storyRepository.findById(storyId).orElseThrow(() -> new NullPointerException("해당하는 일기가 없습니다"));
+        Story story = storyRepository.findById(storyId).orElse(null);
+        if (story == null)
+            throw new StoryHandler(ErrorStatus._STORY_NOT_FOUND);
         writingRepository.deleteByStory(story);
     }
 }
